refactor: move @chart-parts/scene functionality into builder package
