+++ packages/client/builder/package.json
@@ -17,7 +17,12 @@
 	"dependencies": {
 		"@chart-parts/interfaces": "^0.0.20",
 		"lodash": "^4.17.15",
-		"rxjs": "^6.5.3"
+		"rxjs": "^6.5.3",
+		"@chart-parts/scenegraph": "^0.0.20",
+		"@types/d3-format": "^1.3.1",
+		"@types/d3-time-format": "^2.1.0",
+		"d3-format": "^1.4.1",
+		"d3-time-format": "^2.1.1"
 	},
 	"publishConfig": {
 		"access": "public"
