+++ packages/client/builder/src/index.ts
@@ -1,2 +1,3 @@
 export * from './api'
 export * from './spec/defaults'
+export * from './xform'
