+++ packages/client/builder/src/xform/axes/buildAxis.ts
@@ -13,7 +13,7 @@ import { buildMark } from '@chart-parts/scenegraph'
 import { SceneFrame } from '../SceneFrame'
 import { components } from './components'
 import { SGMarkAny } from '../processNode'
-import { AxisSpace } from '../../interfaces'
+import { AxisSpace } from '../interfaces'
 import { getContext } from './getContext'
 import { AxisContext } from './interfaces'
