+++ packages/client/builder/src/xform/createScenegraph.ts
@@ -3,9 +3,9 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import { ChartOptions, SceneNode, DataFrame } from '@chart-parts/interfaces'
-import { ChartOptionsManager } from '@chart-parts/builder'
+import { ChartOptionsManager } from '../api/ChartOptionsManager'
 import { SceneFrame } from './SceneFrame'
-import { GeneratedScene } from '../interfaces'
+import { GeneratedScene } from './interfaces'
 import { processNode } from './processNode'
 
 /**
