+++ packages/client/builder/src/xform/index.ts
@@ -2,6 +2,5 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-
-export * from './scene'
+export * from './createScenegraph'
 export * from './interfaces'
