+++ packages/processing/orchestrator/package.json
@@ -13,7 +13,7 @@
 	},
 	"dependencies": {
 		"@chart-parts/interfaces": "^0.0.20",
-		"@chart-parts/scene": "^0.0.20",
+		"@chart-parts/builder": "^0.0.20",
 		"@chart-parts/scenegraph": "^0.0.20",
 		"@chart-parts/vsvg": "^0.0.20"
 	},
