+++ packages/processing/orchestrator/src/Orchestrator.ts
@@ -13,7 +13,7 @@ import {
 } from '@chart-parts/interfaces'
 import { parseScene } from '@chart-parts/scenegraph'
 import { VirtualSvgConverter } from '@chart-parts/vsvg'
-import { createScenegraph } from '@chart-parts/scene'
+import { createScenegraph } from '@chart-parts/builder'
 
 export class Orchestrator<T> {
 	public constructor(
