--- packages/processing/scene/.npmignore
@@ -1,2 +1,0 @@
-src/
-tsconfig.json
