--- packages/processing/scene/README.md
@@ -1,3 +1,0 @@
-## Purpose
-
-This library processes scene specifications, binding them with data and transforming them into Scenegraph instances.
