--- packages/processing/scene/package.json
@@ -1,27 +1,0 @@
-{
-	"name": "@chart-parts/scene",
-	"description": "Scene-to-Scenegraph Processing",
-	"version": "0.0.20",
-	"main": "dist/cjs/index.js",
-	"module": "dist/esm/index.js",
-	"types": "lib/index.d.ts",
-	"license": "MIT",
-	"scripts": {
-		"clean": "../../../scripts/clean_package.sh",
-		"build": "../../../scripts/build_package.sh",
-		"start": "../../../scripts/watch_package.sh"
-	},
-	"dependencies": {
-		"@chart-parts/builder": "^0.0.20",
-		"@chart-parts/interfaces": "^0.0.20",
-		"@chart-parts/scenegraph": "^0.0.20",
-		"@types/d3-format": "^1.3.1",
-		"@types/d3-time-format": "^2.1.0",
-		"d3-format": "^1.4.1",
-		"d3-time-format": "^2.1.1",
-		"lodash": "^4.17.15"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
-}
