--- packages/processing/scene/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"extends": "../../../tsconfig.json",
-	"compilerOptions": {
-		"lib": ["esnext", "dom"],
-		"outDir": "lib",
-		"composite": true,
-		"rootDir": "./src",
-		"baseUrl": "./src"
-	},
-	"files": ["src/index.ts"],
-	"include": ["./src/**/*.ts*"]
-}
