--- packages/processing/scene/tsconfig.typedoc.json
@@ -1,6 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"composite": false
-	}
-}
